backend/requirements.txt
fastapi
uvicorn
pandas
numpy
torch
scikit-learn
google-generativeai
python-dotenv
pydantic
langchain
sentence-transformers
sqlite3
